requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
mysqlclient==2.1.1
protobuf==4.21.12
Pygments==2.15.1
pytz==2023.3
sqlparse==0.4.4